Made report list clickable and removed view report button
