+++ packages/report-list/components/Report/RemovableReport.jsx
@@ -2,18 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import styled from 'styled-components';
-import {
-  Text,
-} from '@bufferapp/components';
-import CloseIcon from '@bufferapp/components/Icon/Icons/CloseIcon';
-import {
-  Button,
-  VerticalDivider,
-} from '@bufferapp/analyze-shared-components';
+import { Button } from '@bufferapp/components';
 
 import Name from '../Name';
 import Date from '../Date';
 import ReportText from '../ReportText';
+import RemoveButton from '../RemoveButton';
 
 const ButtonContainer = styled.span`
   position: absolute;
@@ -21,32 +15,18 @@ const ButtonContainer = styled.span`
   display: flex;
 `;
 
-
-const DeleteButton = styled(Button)`
-  &:before {
-    width: 1px
-    height: 1px;
-    display: block;
-    content: "";
-  }
-`;
-
 const RemovableReport =
   ({ _id, updated_at, name, small, removeReport, selectReport, showButtons }) =>
-    <ReportText small={small}>
-      <Name name={name} small={small} />
-      { !showButtons && <Date updated_at={updated_at} small={small} /> }
-      { showButtons &&
-        <ButtonContainer>
-          <Button onClick={() => selectReport(_id)}>
-            <Text>View report</Text>
-          </Button>
-          <VerticalDivider />
-          <DeleteButton noStyle onClick={() => removeReport(_id)}>
-            <CloseIcon color="torchRed" />
-          </DeleteButton>
-        </ButtonContainer> }
-    </ReportText>;
+    <Button noStyle fillContainer onClick={() => selectReport(_id)}>
+      <ReportText small={small}>
+        <Name name={name} small={small} />
+        { !showButtons && <Date updated_at={updated_at} small={small} /> }
+        { showButtons &&
+          <ButtonContainer>
+            <RemoveButton _id={_id} removeReport={removeReport} />
+          </ButtonContainer> }
+      </ReportText>
+    </Button>;
 
 RemovableReport.propTypes = {
   _id: PropTypes.string.isRequired,
